@@ -1,64 +1,28 @@
 {
   "name": "admin",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
-  "main": "lib/functions/src/index.js",
-  "dependencies": {
-    "@apollo/client": "^3.5.10",
-    "@sentry/react": "^7.0.0",
-    "@sentry/tracing": "^7.0.0",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.0.1",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.27.0",
-    "date-fns": "^2.29.3",
-    "graphql": "^16.3.0",
-    "i18n-iso-countries": "^7.5.0",
-    "moment": "^2.29.4",
-    "node-polyglot": "^2.4.2",
-    "object.omit": "^3.0.0",
-    "prop-types": "^15.8.1",
-    "ra-input-rich-text": "^4.4.0",
-    "react": "^18.0.0",
-    "react-admin": "^4.4.1",
-    "react-day-picker": "^8.3.5",
-    "react-dom": "^18.0.0",
-    "react-scripts": "5.0.1",
-    "source-map-explorer": "^2.5.2",
-    "typescript": "^4.6.3",
-    "url": "^0.11.0",
-    "web-vitals": "^2.1.4"
-  },
   "scripts": {
-    "analyze": "source-map-explorer 'build/static/js/*.js'",
-    "start": "PORT=5001 react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "admin": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "type-check": "tsc --noEmit"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@apollo/client": "^3.7.15",
+    "axios": "^1.4.0",
+    "graphql": "^16.6.0",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-admin": "^4.9.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.1",
-    "@types/node": "^16.11.26",
-    "@types/react": "^18.0.4",
-    "@types/react-dom": "^18.0.0",
-    "terser-webpack-plugin": "^5.3.2"
+    "@types/node": "^18.16.1",
+    "@types/react": "^18.0.22",
+    "@types/react-dom": "^18.0.7",
+    "@vitejs/plugin-react": "^2.2.0",
+    "typescript": "^4.6.4",
+    "vite": "^3.2.0"
   }
 }
