@@ -1,69 +1,155 @@
-import { AuthProvider } from 'react-admin'
-import { fetcher, poster } from '../lib/fetch.lib'
+import {
+  AuthProvider,
+  HttpError,
+  addRefreshAuthToAuthProvider
+} from 'react-admin'
+import axios from 'axios'
+import { environment } from '../config/environment.config'
+import moment from 'moment'
 
-const authProvider: AuthProvider = {
-  login: async ({ username, password }) => {
-    const [error, data]: any[string] = await poster('/admin/authentication/signin', {
-      email: username,
-      password,
-    })
+export const refreshAuth = () => {
+  const D = localStorage.getItem('login')
 
-    if (error && !data) {
-      return Promise.reject({
-        message: 'ERR_BAD_REQUEST',
+  if (D) {
+    const PastDate = moment(new Date(D))
+    const CurrentDate = moment()
+    var Duration = moment.duration(CurrentDate.diff(PastDate))
+    var Minutes = Duration.asMinutes()
+
+    if (Minutes > 30) {
+      return axios({
+        method: 'get',
+        url: `${environment.apiUrl}/admin/authentication/refresh`,
+        withCredentials: true
       })
+        .then((response) => {
+          if (response.status === 200) {
+            return Promise.resolve()
+          }
+
+          return Promise.reject(
+            new HttpError('Unauthorized', 401, {
+              message: 'Invalid username or password'
+            })
+          )
+        })
+        .catch((e) => {
+          return Promise.reject(
+            new HttpError('Unauthorized', 401, {
+              message: 'Invalid username or password'
+            })
+          )
+        })
     }
 
-    localStorage.setItem('auth', data)
-    // accept all username/password combinations
     return Promise.resolve()
-  },
-  logout: async () => {
-    localStorage.removeItem('auth')
+  }
 
-    return new Promise<void>(async (resolve, _reject) => {
-      const [error, data] = await fetcher('/admin/authentication/signout')
-
-      if (error && !data) {
-        return resolve()
-      }
-
-      // other error code (404, 500, etc): no need to log out
-      return resolve()
+  return Promise.reject(
+    new HttpError('Unauthorized', 401, {
+      message: 'Invalid username or password'
     })
-  },
-  getIdentity: async () => {
-    try {
-      const [error, data] = await fetcher('/admin/authentication')
-      if (error && !data) {
-        localStorage.removeItem('auth')
-        return Promise.reject()
-      }
-
-      localStorage.setItem('auth', data)
-      return Promise.resolve({ id: data.admin.id, fullName: data.admin.name })
-    } catch (error) {
-      return Promise.reject()
-    }
-  },
-  checkError: async (err) => {
-    if (err === 401 || err === 403) {
-      localStorage.removeItem('auth')
-      return Promise.reject()
-    }
-    // other error code (404, 500, etc): no need to log out
-    return Promise.resolve()
-  },
-  checkAuth: async () => {
-    // const [error, data] = await fetcher('/admin/authentication')
-    // if (error && !data) {
-    //   return Promise.reject()
-    // }
-    // return localStorage.getItem('auth') ? Promise.resolve() : Promise.reject()
-
-    return localStorage.getItem('auth') ? Promise.resolve() : Promise.reject({ message: 'login.required' })
-  },
-  getPermissions: () => Promise.resolve('admin'),
+  )
 }
 
-export default authProvider
+export const myAuthProvider: AuthProvider = {
+  login: ({ username, password }) => {
+    return axios({
+      method: 'post',
+      url: `${environment.apiUrl}/admin/authentication/signin`,
+      withCredentials: true,
+      data: {
+        email: username,
+        password
+      }
+    })
+      .then((response) => {
+        if (response.status === 200) {
+          const setDate = new Date()
+
+          localStorage.setItem('login', setDate.toString())
+          return Promise.resolve()
+        }
+
+        return Promise.reject(
+          new HttpError('Unauthorized', 401, {
+            message: 'Invalid username or password'
+          })
+        )
+      })
+      .catch((e) => {
+        return Promise.reject(
+          new HttpError('Unauthorized', 401, {
+            message: 'Invalid username or password'
+          })
+        )
+      })
+  },
+  logout: () => {
+    localStorage.removeItem('login')
+
+    return axios({
+      method: 'get',
+      url: `${environment.apiUrl}/admin/authentication/signout`,
+      withCredentials: true
+    })
+      .then((response) => {
+        if (response.status === 200) {
+          return Promise.resolve()
+        }
+
+        return Promise.resolve()
+      })
+      .catch((e) => {
+        return Promise.resolve()
+      })
+  },
+  checkError: (error) => {
+    const status = error.status
+    if (status === 401 || status === 403) {
+      localStorage.removeItem('login')
+      return Promise.reject()
+    }
+    return Promise.resolve()
+  },
+  checkAuth: () => {
+    return axios({
+      method: 'get',
+      url: `${environment.apiUrl}/admin/authentication`,
+      withCredentials: true
+    })
+      .then((response) => {
+        if (response.status === 200) {
+          return Promise.resolve()
+        }
+
+        return Promise.reject(
+          new HttpError('Unauthorized', 401, {
+            message: 'Invalid username or password'
+          })
+        )
+      })
+      .catch((e) => {
+        return Promise.reject(
+          new HttpError('Unauthorized', 401, {
+            message: 'Invalid username or password'
+          })
+        )
+      })
+    // localStorage.getItem('user') ? Promise.resolve() : Promise.reject(),
+  },
+  getPermissions: () => Promise.resolve(),
+  getIdentity: () => {
+    const persistedUser = localStorage.getItem('user')
+    const user = persistedUser ? JSON.parse(persistedUser) : null
+
+    return Promise.resolve(user)
+  }
+}
+
+export const authProvider = addRefreshAuthToAuthProvider(
+  myAuthProvider,
+  refreshAuth
+)
+
+// export const authProvider = myAuthProvider
