@@ -1,27 +1,5 @@
 import UserIcon from '@mui/icons-material/People'
-import BusinessIcon from '@mui/icons-material/Business'
-import Festival from '@mui/icons-material/Festival'
-import LocationOn from '@mui/icons-material/LocationOn'
-import List from '@mui/icons-material/List'
-import Work from '@mui/icons-material/Work'
-import OpenInNew from '@mui/icons-material/OpenInNew'
-import RemoveRedEye from '@mui/icons-material/RemoveRedEye'
-import Add from '@mui/icons-material/Add'
-import AvTimer from '@mui/icons-material/AvTimer'
-import Article from '@mui/icons-material/Article'
-import CalendarMonth from '@mui/icons-material/CalendarMonth'
 
-export {
-  UserIcon,
-  BusinessIcon,
-  Festival,
-  LocationOn,
-  List,
-  Work,
-  OpenInNew,
-  RemoveRedEye,
-  Add,
-  AvTimer,
-  Article,
-  CalendarMonth,
-}
+export const Icons: {
+  [key: string]: React.ReactElement
+} = { User: <UserIcon /> }
