@@ -26,7 +26,7 @@ const SignIn = async (req: Request, res: Response) => {
         id: user.id,
         email: user.email,
         verfied: user.verfied,
-        Type: user.Type,
+        type: user.type,
         Profile: user.Profile,
         verificationEmail: user.verificationEmail
       }
